Update python code test and functions changed to tools
