+++ requirements.txt
@@ -34,3 +34,4 @@ jq==1.1.2; platform_system == "Linux" or platform_system == "MacOS"
 cryptography==3.3.2; platform_system == "Linux" or platform_system == "MacOS" or platform_system=='Windows'
 urllib3>=1.26.5
 bandit==1.7.0
+git-repo==1.10.3
